feat(pm2): default facility-server to multiprocess deployment
